# backend/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger EmployeeList"
  },
  "host": "localhost:3002",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "oauth2": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey"
    }
  },
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/employees": {
      "post": {
        "security": [
          {
            "oauth2": []
          }
        ],
        "tags": [
          "employees"
        ],
        "summary": "Add a new employee to the list",
        "operationId": "createEmployee",
        "consumes": "application/json",
        "produces": "application/json",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Employee object that needs to be added to the list",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "empName": {
                  "type": "string"
                },
                "empActive": {
                  "type": "boolean"
                },
                "empDepartment": {
                  "type": "integer"
                }
              },
              "example": {
                "empName": "Uncle Bob",
                "empActive": 1,
                "empDepartment": 1
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "You are not authenticated"
          },
          "405": {
            "description": "Invalid input"
          },
          "UnauthorizedError": {
            "description": "Access token is missing or invalid"
          }
        }
      }
    },
...
  },
...
}